Move `WindowsAssociationManager` to `osu.Desktop`
